feat: 將 headers 部分，移至 lib.js 管理
