+++ server.js
@@ -1,21 +1,10 @@
 const http = require('http');
 const { v4: uuidv4 } = require('uuid');
-const errHandle = require('./errorHandle');
+const { headers } = require('./libs');
+const errorHandle = require('./errorHandle');
 const todos = [];
 
 const requestListener = (req, res) => {
-  const headers = {
-    'Access-Control-Allow-Headers': 'Content-Type, Authorization, Content-Length, X-Requested-With',
-    'Access-Control-Allow-Origin': '*',
-    'Access-Control-Allow-Methods': 'PATCH, POST, GET,OPTIONS,DELETE',
-    'Content-Type': 'application/json'
-  }
-  let body = "";
-
-  req.on('data', chunk => {
-    body += chunk;
-  })
-
   if (req.url == "/todos" && req.method == "GET") {
     // getTodo.js
   } else if (req.url == "/todos" && req.method == "POST") {
